Refactor preset loading and fix UnoCSS variant handling
