+++ packages/utils/src/index.ts
@@ -40,8 +40,8 @@ export * from './key-case'
 export * from './match-run'
 export * from './number-to'
 export * from './once'
-export * from './pascal-case'
 export * from './parse-css-length'
+export * from './pascal-case'
 export * from './path'
 export * from './pipe'
 export * from './prev-args'
